Enough work tonight.
Wrapping my brain around the job suspension. I'm wondering if we're supposed to use multithreading or something, but now I think we'll need to add a notify RMI function to Client that will allow the Server to command it to download something. We'll also need to make sure to add proper suspended download handling in the client side.

Bulk of the remainder after the job suspension stuff is getting the client to work along with chmod and emacs stuff. Thankfully the hard work around text file handling is written.