Major changes happened in Client. Details follow.
I added a time-out logic to the end of userPrompt() that would trigger a "please use a text editor to continue" prompt in the case that the emacs exec and waitFor completed in less than 100ms. This allows a local terminal to continue immediately after emacs is closed (after being open for at least one-tenth of a second).

This is a two-stage process. This commit is stage one and has not been tested:

In the userPrompt switch case for Release_Ownership, handling of the client requesting the same file or a different file when the server's writeback function has been called has been handled. If the client wants to access the same file that was just uploaded to the server via a syncing writeBack function call the program will notify that the server has forced their access to the file to be read-only mode.
If trying to access another file, the download is now no-longer ignored as before and will perform the necessary steps to access the file.

Stage two will be to enforce waiting by the client for a file it does not have access to (possibly waiting for a writeBack sync to complete).