Even more minor polishing for client.
Added consistency between the two uses of pushFile() where clientState and accessMode and hasOwnership are set to Invalid/false.

Still worried about the client going into an infinite wait due to a poorly-timed invalidate() call. The spin wait is for failed downloads (and thus must wait for a resume()), but invalidates and failed downloads don't have different signatures to go off of.

...

Thinking it over, an invalidate call would have to happen while the userPrompt thread is going through the state switch. Because at any time an invalidate would simply happen before the switch and be handled just fine... if an invalidate happened while using emacs on the local machine, it would likely cause emacs to throw up an error due to "chmod 000"...

To wait on a bad signal it would either
A) get an invalidate during actual code execution (a window of perhaps <5ms) or
B) it arrives within the spin wait() cycle after a resume() call.