Complex changes to thread sync.
Client:
Implemented synchronized wait() and notify() between the prompt and main thread.

Server:
Changed the relationship between the OC and WS queues, instead of knowingly resuming an OC queue member into a failed download attempt, it instead simply puts it into the WS queue and removes it from the OC queue to save the trip.

That, and various cout format changes for cleaner output.